Avoid Saas mixed declaration warnings
The current version of the bootstrap gem (v5.3.3) is triggering a bunch
of these deprecation warnings [1] in the build.

I reported the problem in this issue [2], but it has been marked as a
duplication of this issue [3]. It's not currently fixed, but it will
supposedly be fixed [4] in the next release of bootstrap.

Note that the relevant sass change seems to have been introduced in
v1.77.7 of sass-embedded. See this changelog entry [5] for details.

[1]: https://sass-lang.com/d/mixed-decls
[2]: twbs/bootstrap-rubygem#285
[3]: twbs/bootstrap#40621
[4]: twbs/bootstrap-rubygem#285 (comment)
[5]: https://github.com/sass/embedded-host-node/blob/main/CHANGELOG.md#1777